+++ src/Rule/RuleInterface.php
@@ -15,10 +15,11 @@ interface RuleInterface
      * Check whether a password is in compliance with the rule.
      *
      * @param Password|string $password Password to check.
+     * @param int|null $weight Don't consider constraints with lower weights.
      * @return bool Whether the password is in compliance with the rule.
      * @throws TestException If an error occurred while testing the rule.
      */
-    public function test($password): bool;
+    public function test($password, ?int $weight = 1): bool;
 
     /**
      * Enforce that a password is in compliance with the rule.