+++ tests/Rule/LowerCaseTest.php
@@ -183,6 +183,16 @@ final class LowerCaseTest extends TestCase
     }
 
     /**
+     * @covers ::test
+     */
+    public function testRuleIsSatisfiedWhenConstraintWeightIsLowerThanTestWeight(): void
+    {
+        $rule = new LowerCase(0, 3, 1);
+
+        self::assertTrue($rule->test('foo bar', 2));
+    }
+
+    /**
      * @covers ::enforce
      */
     public function testEnforceDoesNotThrowExceptionWhenRuleIsSatisfied(): void